--- a/rtree/box.go
+++ b/rtree/box.go
@@ -1,7 +1,5 @@
 package rtree
 
-import "math"
-
 // Box is an axis-aligned bounding box.
 type Box struct {
 	MinX, MinY, MaxX, MaxY float64
@@ -19,10 +17,10 @@ func calculateBound(n *node) Box {
 // combine gives the smallest bounding box containing both box1 and box2.
 func combine(box1, box2 Box) Box {
 	return Box{
-		MinX: math.Min(box1.MinX, box2.MinX),
-		MinY: math.Min(box1.MinY, box2.MinY),
-		MaxX: math.Max(box1.MaxX, box2.MaxX),
-		MaxY: math.Max(box1.MaxY, box2.MaxY),
+		MinX: fastMin(box1.MinX, box2.MinX),
+		MinY: fastMin(box1.MinY, box2.MinY),
+		MaxX: fastMax(box1.MaxX, box2.MaxX),
+		MaxY: fastMax(box1.MaxY, box2.MaxY),
 	}
 }
 
@@ -43,7 +41,7 @@ func overlap(box1, box2 Box) bool {
 }
 
 func squaredEuclideanDistance(b1, b2 Box) float64 {
-	dx := math.Max(0, math.Max(b1.MinX-b2.MaxX, b2.MinX-b1.MaxX))
-	dy := math.Max(0, math.Max(b1.MinY-b2.MaxY, b2.MinY-b1.MaxY))
+	dx := fastMax(0, fastMax(b1.MinX-b2.MaxX, b2.MinX-b1.MaxX))
+	dy := fastMax(0, fastMax(b1.MinY-b2.MaxY, b2.MinY-b1.MaxY))
 	return dx*dx + dy*dy
 }