## Description

Use fastMin and fastMax throughout the codebase indiscriminately.

The intention was to get Union/Intersection etc to run a bit faster (math.Min and math.Max were combining to 2.5% of the runtime in the CPU profile). This should theoretically have occurred in this PR, but doesn't show up in the benchmark results because the benchmark variance is higher than that.

_However_, significant performance improvements are seen across the board for a lot of other benchmarks, so it's definitely worthwhile.

The only behavioural difference is when `NaN` is involved. However, all bets are off the table anyway if `NaN` becomes involved (so it doesn't really matter if `NaN`s are propagated in the standard floating point way or not).

## Check List

Have you:

- Added unit tests? No, uses existing.

- Add cmprefimpl tests? (if appropriate?) No, uses existing.

## Related Issue

N/A

## Benchmark Results

```
COMPARISON
name                                                        old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0-4                                              1.65ns ±12%    1.58ns ± 4%   -4.15%  (p=0.011 n=13+14)
LineEnvelope/1-4                                              1.05ns ± 8%    1.04ns ± 4%     ~     (p=0.308 n=14+14)
LineEnvelope/2-4                                              1.31ns ± 5%    1.31ns ± 9%     ~     (p=0.488 n=15+15)
LineEnvelope/3-4                                              1.19ns ± 6%    1.20ns ± 8%     ~     (p=0.991 n=14+15)
MarshalWKB/polygon/n=10-4                                      199ns ±26%     195ns ±19%     ~     (p=0.493 n=13+14)
MarshalWKB/polygon/n=100-4                                    701ns ±116%     827ns ±49%     ~     (p=0.068 n=13+14)
MarshalWKB/polygon/n=1000-4                                  7.27µs ±151%   6.66µs ±113%     ~     (p=0.928 n=15+13)
MarshalWKB/polygon/n=10000-4                                  37.7µs ±48%   56.5µs ±150%     ~     (p=0.399 n=12+15)
UnmarshalWKB/polygon/n=10-4                                    313ns ±10%     312ns ±12%     ~     (p=0.742 n=15+13)
UnmarshalWKB/polygon/n=100-4                                   755ns ±29%     849ns ±81%     ~     (p=0.553 n=13+13)
UnmarshalWKB/polygon/n=1000-4                                 6.36µs ±93%   7.19µs ±130%     ~     (p=0.652 n=15+14)
UnmarshalWKB/polygon/n=10000-4                               68.1µs ±103%   85.4µs ±116%     ~     (p=0.567 n=15+15)
IntersectsLineStringWithLineString/n=10-4                     1.61µs ±33%    1.44µs ±11%  -11.06%  (p=0.001 n=12+14)
IntersectsLineStringWithLineString/n=100-4                    21.3µs ±20%    21.8µs ±31%     ~     (p=0.860 n=12+14)
IntersectsLineStringWithLineString/n=1000-4                   317µs ±103%    236µs ±118%  -25.39%  (p=0.009 n=14+13)
IntersectsLineStringWithLineString/n=10000-4                  3.23ms ±21%    3.07ms ±13%     ~     (p=0.252 n=15+14)
IntersectsMultiPointWithMultiPoint/n=20-4                     1.31µs ± 8%    1.31µs ± 7%     ~     (p=0.940 n=15+14)
IntersectsMultiPointWithMultiPoint/n=200-4                    14.1µs ±12%    14.2µs ±10%     ~     (p=0.910 n=14+14)
IntersectsMultiPointWithMultiPoint/n=2000-4                    142µs ±11%     139µs ± 4%     ~     (p=0.650 n=14+13)
IntersectsMultiPointWithMultiPoint/n=20000-4                  1.51ms ± 9%    1.53ms ±10%     ~     (p=0.400 n=14+15)
LineStringIsSimpleZigZag/10-4                                 1.92µs ± 9%    1.87µs ± 9%     ~     (p=0.137 n=14+13)
LineStringIsSimpleZigZag/100-4                                31.0µs ± 8%    29.3µs ± 9%   -5.38%  (p=0.001 n=14+12)
LineStringIsSimpleZigZag/1000-4                                378µs ± 6%     370µs ±15%     ~     (p=0.114 n=12+15)
LineStringIsSimpleZigZag/10000-4                              5.08ms ±10%    4.79ms ±11%   -5.69%  (p=0.003 n=14+13)
PolygonSingleRingValidation/n=10-4                            2.59µs ±10%    2.41µs ±21%   -7.18%  (p=0.009 n=13+13)
PolygonSingleRingValidation/n=100-4                           36.1µs ±14%    31.7µs ±13%  -12.23%  (p=0.000 n=12+14)
PolygonSingleRingValidation/n=1000-4                           433µs ±14%     381µs ± 5%  -12.09%  (p=0.000 n=13+13)
PolygonSingleRingValidation/n=10000-4                         5.23ms ± 9%    5.02ms ±11%   -4.13%  (p=0.046 n=13+15)
PolygonMultipleRingsValidation/n=4-4                          6.79µs ± 8%    6.80µs ±16%     ~     (p=0.880 n=15+14)
PolygonMultipleRingsValidation/n=36-4                         64.5µs ± 9%    57.6µs ±11%  -10.62%  (p=0.000 n=13+14)
PolygonMultipleRingsValidation/n=400-4                         868µs ±12%     787µs ±12%   -9.37%  (p=0.001 n=13+14)
PolygonMultipleRingsValidation/n=4096-4                       10.8ms ±16%     9.1ms ±13%  -15.57%  (p=0.000 n=13+14)
PolygonZigZagRingsValidation/n=10-4                           11.8µs ±11%    11.1µs ±21%   -5.84%  (p=0.009 n=14+14)
PolygonZigZagRingsValidation/n=100-4                           134µs ±12%     129µs ±12%     ~     (p=0.058 n=13+15)
PolygonZigZagRingsValidation/n=1000-4                         1.52ms ±10%    1.44ms ±10%   -5.26%  (p=0.008 n=15+15)
PolygonZigZagRingsValidation/n=10000-4                        19.4ms ±10%    19.4ms ±13%     ~     (p=0.742 n=12+14)
PolygonAnnulusValidation/n=10-4                               3.82µs ± 8%    3.70µs ± 8%     ~     (p=0.139 n=13+13)
PolygonAnnulusValidation/n=100-4                              33.8µs ± 5%    32.5µs ±20%   -3.80%  (p=0.014 n=13+14)
PolygonAnnulusValidation/n=1000-4                              595µs ±14%     550µs ±14%   -7.65%  (p=0.001 n=14+14)
PolygonAnnulusValidation/n=10000-4                            7.09ms ±17%    6.47ms ±13%   -8.77%  (p=0.002 n=14+12)
MultipolygonValidation/n=1-4                                   348ns ± 9%     300ns ±14%  -13.77%  (p=0.000 n=14+13)
MultipolygonValidation/n=4-4                                   909ns ± 9%     713ns ±15%  -21.60%  (p=0.000 n=15+12)
MultipolygonValidation/n=16-4                                 5.51µs ± 6%    4.03µs ± 5%  -26.88%  (p=0.000 n=15+12)
MultipolygonValidation/n=64-4                                 33.5µs ±11%    22.6µs ±19%  -32.36%  (p=0.000 n=13+15)
MultipolygonValidation/n=256-4                                 194µs ± 9%     158µs ± 8%  -18.37%  (p=0.000 n=13+14)
MultipolygonValidation/n=1024-4                                976µs ±17%     767µs ±15%  -21.42%  (p=0.000 n=14+15)
MultiPolygonTwoCircles/n=10-4                                 4.10µs ±16%    3.53µs ±26%  -13.97%  (p=0.000 n=12+12)
MultiPolygonTwoCircles/n=100-4                                42.6µs ±15%    36.1µs ±11%  -15.39%  (p=0.000 n=13+14)
MultiPolygonTwoCircles/n=1000-4                                433µs ± 4%     409µs ±16%   -5.71%  (p=0.000 n=13+12)
MultiPolygonTwoCircles/n=10000-4                              6.35ms ±15%    5.79ms ±12%   -8.82%  (p=0.002 n=13+14)
MultiPolygonMultipleTouchingPoints/n=1-4                      4.90µs ±16%    4.59µs ±11%   -6.25%  (p=0.007 n=13+13)
MultiPolygonMultipleTouchingPoints/n=10-4                     39.6µs ± 9%    38.3µs ±25%     ~     (p=0.105 n=14+13)
MultiPolygonMultipleTouchingPoints/n=100-4                     470µs ±16%     422µs ± 3%  -10.14%  (p=0.000 n=15+13)
MultiPolygonMultipleTouchingPoints/n=1000-4                   5.29ms ± 9%    4.99ms ± 3%   -5.72%  (p=0.000 n=13+14)
Intersection/n=10-4                                           55.6µs ±13%    54.2µs ± 6%     ~     (p=0.220 n=14+13)
Intersection/n=100-4                                           299µs ±12%     295µs ±13%     ~     (p=0.277 n=12+15)
Intersection/n=1000-4                                         2.81ms ± 6%    2.70ms ±14%   -4.08%  (p=0.016 n=13+13)
Intersection/n=10000-4                                        30.4ms ±18%    29.1ms ±20%     ~     (p=0.137 n=13+12)
WKTParsing/point-4                                            1.75µs ±11%    1.76µs ±17%     ~     (p=0.905 n=15+12)
DistancePolygonToPolygonOrdering/n=100_swap=false-4           71.4µs ±14%    49.7µs ±13%  -30.49%  (p=0.000 n=14+13)
DistancePolygonToPolygonOrdering/n=100_swap=true-4            70.5µs ±13%    49.3µs ± 8%  -30.10%  (p=0.000 n=14+15)
DistancePolygonToPolygonOrdering/n=1000_swap=false-4          1.20ms ± 5%    0.75ms ± 5%  -37.08%  (p=0.000 n=13+13)
DistancePolygonToPolygonOrdering/n=1000_swap=true-4           1.24ms ±14%    0.74ms ± 3%  -40.21%  (p=0.000 n=14+12)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-4     5.59µs ±13%    5.57µs ±12%     ~     (p=0.756 n=14+13)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-4      5.63µs ±12%    5.48µs ±10%     ~     (p=0.123 n=13+12)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-4    73.9µs ±71%    60.2µs ±22%     ~     (p=0.062 n=14+14)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-4     61.1µs ±27%    58.5µs ±10%     ~     (p=0.259 n=13+14)
MultiLineStringIsSimpleManyLineStrings/n=100-4                55.0µs ±25%    51.7µs ±16%     ~     (p=0.085 n=14+14)
MultiLineStringIsSimpleManyLineStrings/n=1000-4                568µs ±14%     578µs ±22%     ~     (p=0.829 n=15+12)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                                           47.2µs ±10%    46.5µs ± 6%     ~     (p=0.561 n=14+15)
Intersection/n=100-4                                          92.1µs ±13%    94.2µs ± 9%     ~     (p=0.254 n=13+15)
Intersection/n=1000-4                                          391µs ± 7%     394µs ±18%     ~     (p=0.946 n=14+14)
Intersection/n=10000-4                                        3.65ms ±13%    3.44ms ± 5%   -5.66%  (p=0.016 n=13+13)
NoOp/n=10-4                                                   3.80µs ± 3%    3.81µs ± 6%     ~     (p=0.971 n=13+14)
NoOp/n=100-4                                                  12.6µs ±15%    12.7µs ±13%     ~     (p=0.715 n=14+15)
NoOp/n=1000-4                                                 86.9µs ± 5%    90.5µs ±11%   +4.21%  (p=0.041 n=13+15)
NoOp/n=10000-4                                                1.01ms ±34%    0.92ms ±21%     ~     (p=0.156 n=15+13)
pkg:github.com/peterstace/simplefeatures/rtree goos:linux goarch:amd64
Delete/n=100-4                                                30.5µs ± 7%    19.7µs ± 4%  -35.61%  (p=0.000 n=14+15)
Delete/n=1000-4                                                697µs ± 3%     651µs ± 6%   -6.53%  (p=0.000 n=12+14)
Delete/n=10000-4                                              31.1ms ± 2%    30.9ms ±10%     ~     (p=0.067 n=12+14)
Bulk/n=10-4                                                   970ns ±106%     818ns ±15%  -15.73%  (p=0.037 n=12+12)
Bulk/n=100-4                                                  16.2µs ±45%    14.2µs ±30%  -12.69%  (p=0.029 n=15+13)
Bulk/n=1000-4                                                  223µs ±12%     222µs ±10%     ~     (p=1.000 n=15+12)
Bulk/n=10000-4                                                3.04ms ± 6%    3.13ms ±12%     ~     (p=0.176 n=14+12)
Bulk/n=100000-4                                               36.3ms ±13%    35.8ms ±10%     ~     (p=0.541 n=14+14)
Insert/n=10-4                                                 1.42µs ±15%    1.18µs ±25%  -17.01%  (p=0.000 n=13+13)
Insert/n=100-4                                                33.0µs ± 7%    20.7µs ±12%  -37.10%  (p=0.000 n=12+15)
Insert/n=1000-4                                                544µs ± 9%     442µs ±10%  -18.71%  (p=0.000 n=14+15)
Insert/n=10000-4                                              7.60ms ± 9%    5.36ms ±10%  -29.45%  (p=0.000 n=15+14)
Insert/n=100000-4                                             92.3ms ± 3%    59.8ms ± 2%  -35.16%  (p=0.000 n=12+12)
RangeSearch/n=10-4                                            15.0ns ± 4%    14.7ns ± 2%   -1.92%  (p=0.020 n=15+13)
RangeSearch/n=100-4                                           58.4ns ± 7%    58.2ns ± 3%     ~     (p=0.777 n=14+14)
RangeSearch/n=1000-4                                           213ns ± 2%     213ns ± 2%     ~     (p=0.759 n=13+14)
RangeSearch/n=10000-4                                          732ns ± 2%     734ns ± 4%     ~     (p=0.898 n=12+12)
RangeSearch/n=100000-4                                        7.19µs ± 6%    7.25µs ± 9%     ~     (p=0.723 n=15+14)

name                                                        old alloc/op   new alloc/op   delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0-4                                               0.00B          0.00B          ~     (all equal)
LineEnvelope/1-4                                               0.00B          0.00B          ~     (all equal)
LineEnvelope/2-4                                               0.00B          0.00B          ~     (all equal)
LineEnvelope/3-4                                               0.00B          0.00B          ~     (all equal)
MarshalWKB/polygon/n=10-4                                       232B ± 0%      232B ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100-4                                    1.83kB ± 0%    1.83kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000-4                                   16.4kB ± 0%    16.4kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000-4                                   164kB ± 0%     164kB ± 0%     ~     (p=1.000 n=15+15)
UnmarshalWKB/polygon/n=10-4                                     284B ± 0%      284B ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100-4                                  1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000-4                                 16.5kB ± 0%    16.5kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000-4                                 164kB ± 0%     164kB ± 0%     ~     (p=0.056 n=15+12)
IntersectsLineStringWithLineString/n=10-4                     2.42kB ± 0%    2.42kB ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=100-4                    30.4kB ± 0%    30.4kB ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=1000-4                    205kB ± 0%     205kB ± 0%     ~     (p=0.582 n=14+15)
IntersectsLineStringWithLineString/n=10000-4                  2.63MB ± 0%    2.63MB ± 0%   -0.00%  (p=0.012 n=14+14)
IntersectsMultiPointWithMultiPoint/n=20-4                       324B ± 0%      325B ± 0%     ~     (p=1.000 n=15+15)
IntersectsMultiPointWithMultiPoint/n=200-4                    3.08kB ± 0%    3.08kB ± 0%     ~     (p=0.174 n=15+15)
IntersectsMultiPointWithMultiPoint/n=2000-4                   49.3kB ± 0%    49.3kB ± 0%     ~     (p=0.592 n=15+14)
IntersectsMultiPointWithMultiPoint/n=20000-4                   339kB ± 0%     339kB ± 0%     ~     (p=0.735 n=15+15)
LineStringIsSimpleZigZag/10-4                                 1.84kB ± 0%    1.84kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/100-4                                24.0kB ± 0%    24.0kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/1000-4                                139kB ± 0%     139kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/10000-4                              1.97MB ± 0%    1.97MB ± 0%     ~     (p=0.605 n=14+13)
PolygonSingleRingValidation/n=10-4                            2.19kB ± 0%    2.19kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=100-4                           24.3kB ± 0%    24.3kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=1000-4                           140kB ± 0%     140kB ± 0%     ~     (p=0.833 n=15+15)
PolygonSingleRingValidation/n=10000-4                         1.97MB ± 0%    1.97MB ± 0%     ~     (p=0.671 n=15+15)
PolygonMultipleRingsValidation/n=4-4                          6.22kB ± 0%    6.22kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=36-4                         50.4kB ± 0%    50.4kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=400-4                         553kB ± 0%     553kB ± 0%     ~     (p=0.582 n=14+15)
PolygonMultipleRingsValidation/n=4096-4                       5.67MB ± 0%    5.67MB ± 0%     ~     (p=0.838 n=14+15)
PolygonZigZagRingsValidation/n=10-4                           9.38kB ± 0%    9.38kB ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=100-4                          87.8kB ± 0%    87.8kB ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=1000-4                          551kB ± 0%     551kB ± 0%     ~     (p=0.572 n=15+15)
PolygonZigZagRingsValidation/n=10000-4                        7.24MB ± 0%    7.24MB ± 0%     ~     (p=0.273 n=14+13)
PolygonAnnulusValidation/n=10-4                               3.94kB ± 0%    3.94kB ± 0%     ~     (all equal)
PolygonAnnulusValidation/n=100-4                              28.3kB ± 0%    28.3kB ± 0%     ~     (all equal)
PolygonAnnulusValidation/n=1000-4                              379kB ± 0%     379kB ± 0%   +0.00%  (p=0.035 n=14+13)
PolygonAnnulusValidation/n=10000-4                            3.89MB ± 0%    3.89MB ± 0%   -0.00%  (p=0.032 n=15+14)
MultipolygonValidation/n=1-4                                    385B ± 0%      385B ± 0%     ~     (all equal)
MultipolygonValidation/n=4-4                                    676B ± 0%      676B ± 0%     ~     (all equal)
MultipolygonValidation/n=16-4                                 3.86kB ± 0%    3.86kB ± 0%     ~     (all equal)
MultipolygonValidation/n=64-4                                 15.4kB ± 0%    15.4kB ± 0%     ~     (all equal)
MultipolygonValidation/n=256-4                                63.1kB ± 0%    63.1kB ± 0%     ~     (all equal)
MultipolygonValidation/n=1024-4                                259kB ± 0%     259kB ± 0%     ~     (p=0.115 n=14+15)
MultiPolygonTwoCircles/n=10-4                                 4.99kB ± 0%    4.99kB ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=100-4                                55.0kB ± 0%    55.0kB ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=1000-4                                345kB ± 0%     345kB ± 0%     ~     (p=0.474 n=15+15)
MultiPolygonTwoCircles/n=10000-4                              4.60MB ± 0%    4.60MB ± 0%     ~     (p=0.586 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1-4                      3.95kB ± 0%    3.95kB ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=10-4                     22.6kB ± 0%    22.6kB ± 0%     ~     (p=0.575 n=15+15)
MultiPolygonMultipleTouchingPoints/n=100-4                     172kB ± 0%     172kB ± 0%     ~     (p=0.348 n=14+15)
MultiPolygonMultipleTouchingPoints/n=1000-4                   2.04MB ± 0%    2.04MB ± 0%     ~     (p=0.467 n=15+15)
Intersection/n=10-4                                           29.2kB ± 0%    29.2kB ± 0%     ~     (p=0.158 n=13+15)
Intersection/n=100-4                                           155kB ± 0%     155kB ± 0%     ~     (p=0.330 n=15+15)
Intersection/n=1000-4                                         1.88MB ± 0%    1.88MB ± 0%     ~     (p=0.574 n=15+15)
Intersection/n=10000-4                                        19.3MB ± 0%    19.3MB ± 0%     ~     (p=0.967 n=15+15)
WKTParsing/point-4                                            1.93kB ± 0%    1.93kB ± 0%     ~     (all equal)
DistancePolygonToPolygonOrdering/n=100_swap=false-4           40.7kB ± 0%    40.7kB ± 0%     ~     (all equal)
DistancePolygonToPolygonOrdering/n=100_swap=true-4            40.7kB ± 0%    40.7kB ± 0%     ~     (all equal)
DistancePolygonToPolygonOrdering/n=1000_swap=false-4           369kB ± 0%     369kB ± 0%   -0.02%  (p=0.000 n=15+15)
DistancePolygonToPolygonOrdering/n=1000_swap=true-4            369kB ± 0%     369kB ± 0%   -0.02%  (p=0.000 n=14+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-4     5.52kB ± 0%    5.52kB ± 0%     ~     (all equal)
…